+++ packages/core/acme-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v5.4.0
+
+* `added` Directory URLs for [Google](https://cloud.google.com/certificate-manager/docs/overview) ACME provider
+* `fixed` Invalidate ACME directory cache after 24 hours
+
 ## v5.3.1 (2024-05-22)
 
 * `fixed` Allow `client.auto()` being called with an empty CSR common name